delete component table
